test(traces): keep cost source helper internal
