+++ web/src/features/traces/components/ObservationDetailView/components/ObservationDetailViewHeader/costSource.ts
@@ -1,0 +1,13 @@
+import type { CostSource } from "@/src/features/traces/components/BreakdownTooltip";
+
+export function resolveObservationCostSource({
+  hasSubtreeMetrics,
+  hasProvidedCostDetails,
+}: {
+  hasSubtreeMetrics: boolean;
+  hasProvidedCostDetails: boolean;
+}): CostSource | undefined {
+  if (hasSubtreeMetrics) return undefined;
+
+  return hasProvidedCostDetails ? "provided" : "calculated";
+}
